Final sprint solution Изменения после проверки: Изменил класс BaseHttpHandler Создал enum для статусов и методов Обработал пересечение Разъединил классы из одного файла
