+++ src/status/HttpMethod.java
@@ -1,0 +1,7 @@
+package status;
+
+public enum HttpMethod {
+    GET,
+    POST,
+    DELETE
+}
