+++ src/status/HttpStatus.java
@@ -1,0 +1,21 @@
+package status;
+
+public enum HttpStatus {
+    OK(200),
+    CREATED(201),
+    BAD_REQUEST(400),
+    NOT_FOUND(404),
+    METHOD_NOT_ALLOWED(405),
+    NOT_ACCEPTABLE(406),
+    INTERNAL_SERVER_ERROR(500);
+
+    private final int code;
+
+    HttpStatus(int code) {
+        this.code = code;
+    }
+
+    public int getCode() {
+        return code;
+    }
+}
