+++ test/HistoryManagerTest.java
@@ -1,6 +1,5 @@
 package test;
 
-import history.HistoryManager;
 import tasks.Epic;
 import tasks.Subtask;
 import tasks.Task;
