+++ test/InMemoryTaskManagerTest.java
@@ -1,7 +1,6 @@
 package test;
 
 import manager.InMemoryTaskManager;
-import manager.TaskManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tasks.Epic;
