+++ test/SubtaskEndpointsTest.java
@@ -2,13 +2,16 @@ package test;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import http.DurationAdapter;
 import http.HttpTaskServer;
+import http.LocalDateTimeAdapter;
 import manager.Managers;
 import manager.TaskManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import status.HttpStatus;
 import tasks.Epic;
 import tasks.Subtask;
 
@@ -28,8 +31,8 @@ public class SubtaskEndpointsTest {
     private final Subtask subtask11 = new Subtask("subtask", "description", 1);
     private final TaskManager taskManager = Managers.getDefault();
     private static final Gson gson = new GsonBuilder()
-            .registerTypeAdapter(Duration.class, new HttpTaskServer.DurationAdapter())
-            .registerTypeAdapter(LocalDateTime.class, new HttpTaskServer.LocalDateTimeAdapter())
+            .registerTypeAdapter(Duration.class, new DurationAdapter())
+            .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
             .create();
 
     @BeforeEach
@@ -52,7 +55,7 @@ public class SubtaskEndpointsTest {
                 .uri(uri)
                 .build();
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response.statusCode(), "Not right response");
     }
 
     @Test
@@ -64,7 +67,7 @@ public class SubtaskEndpointsTest {
                 .uri(uri1)
                 .build();
         HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response1.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response1.statusCode(), "Not right response");
         URI uri2 = URI.create(DEFAULT_SUBTASK_URI);
         HttpRequest request2 = HttpRequest.newBuilder()
                 .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask11)))
@@ -72,14 +75,14 @@ public class SubtaskEndpointsTest {
                 .uri(uri2)
                 .build();
         HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response2.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response2.statusCode(), "Not right response");
         URI uri3 = URI.create(DEFAULT_SUBTASK_URI);
         HttpRequest request3 = HttpRequest.newBuilder()
                 .GET()
                 .uri(uri3)
                 .build();
         HttpResponse<String> response3 = client.send(request3, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response3.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response3.statusCode(), "Not right response");
     }
 
     @Test
@@ -91,7 +94,7 @@ public class SubtaskEndpointsTest {
                 .uri(uri1)
                 .build();
         HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response1.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response1.statusCode(), "Not right response");
         URI uri2 = URI.create(DEFAULT_SUBTASK_URI);
         HttpRequest request2 = HttpRequest.newBuilder()
                 .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask11)))
@@ -99,7 +102,7 @@ public class SubtaskEndpointsTest {
                 .uri(uri2)
                 .build();
         HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response2.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.OK.getCode(), response2.statusCode(), "Not right response");
         URI uri3 = URI.create(DEFAULT_SUBTASK_URI + "/1");
         HttpRequest request3 = HttpRequest.newBuilder()
                 .DELETE()
@@ -107,6 +110,6 @@ public class SubtaskEndpointsTest {
                 .uri(uri2)
                 .build();
         HttpResponse<String> response3 = client.send(request3, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(201, response3.statusCode(), "Not right response");
+        Assertions.assertEquals(HttpStatus.CREATED.getCode(), response3.statusCode(), "Not right response");
     }
 }
